Make works

+++ tools/make.py
@@ -422,25 +422,25 @@ def cleanup_optionals(mod):
 
             try:
                 file_name = "{}{}.pbo".format(pbo_name_prefix,dir_name)
-                folder= "@{}{}".format(pbo_name_prefix,dir_name)
+                folder = "@{}{}".format(pbo_name_prefix,dir_name)
                 src_file_path = os.path.join(release_dir, project, "addons", file_name)
-                dst_file_path = os.path.join(release_dir, project, "optionals",folder,"addons",file_name)
-                os.remove(dst_file_path);
+                dst_file_path = os.path.join(release_dir, project, "optionals", folder, "addons", file_name)
 
                 sigFile_name = "{}.{}.bisign".format(file_name,key_name)
                 src_sig_path = os.path.join(release_dir, project, "addons", sigFile_name)
-                dst_sig_path = os.path.join(release_dir, project, "optionals",folder,"addons", sigFile_name)
-                os.remove(dst_sig_path);
-
+                dst_sig_path = os.path.join(release_dir, project, "optionals", folder, "addons", sigFile_name)
 
                 if (os.path.isfile(src_file_path)):
                     #print("Preserving {}".format(file_name))
+                    if (os.path.isfile(dst_file_path)):
+                        os.remove(dst_file_path)
                     os.renames(src_file_path,dst_file_path)
                 if (os.path.isfile(src_sig_path)):
                     #print("Preserving {}".format(sigFile_name))
+                    if (os.path.isfile(dst_sig_path)):
+                        os.remove(dst_sig_path)
                     os.renames(src_sig_path,dst_sig_path)
-            except FileExistsError as e:
-                print_yellow(e)
+            except FileExistsError:
                 print_error("{} already exists".format(file_name))
                 continue
             shutil.rmtree(destination)
